sql: rename forEachTableInMultiRegionDatabase to forEachMutableTableInDatabase
In preparation for the zone config check change, this call must now
allow non-multi-region databases too.

Release note: None
